requirements.txt
Flask==2.1.2
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.23
Flask-Migrate==3.0.1
Flask-CORS==3.0.10
python-dotenv==0.19.0
psycopg2-binary==2.9.1
kafka-python==2.0.2
